Логика взаимодействия с формой перенесена в отдельный hook
